requirement.txt
# environtment handling
python-dotenv==1.0.0

# data handling
numpy==1.26.2
pandas==2.1.3
fsspec==2023.12.1
openpyxl==3.1.2

# visualization
matplotlib==3.8.2
streamlit==1.13.0

# cross-validation
scikit-learn==1.3.2

# modeling and augmentation
tensorflow==2.10.1

# augmentation using clahe
tf_clahe=0.1.0